backend/requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# HTML 解析
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Word 文档处理
python-docx==1.1.0

# 工具库
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2  # AI API 调用


